+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+* fixed pseudo class issue and updated test.
+
 ## [2.0.3] - 2016-08-10
 
 * Fixed issue with removing place holders.